HealthBridge_PY_FlaskWebProject1/requirements.txt
azure-common			
azure-core				
azure-functions			
azure-identity			
azure-keyvault-secrets	
azure-storage-blob		
certifi					
cffi					
charset-normalizer		
click					
colorama				
cryptography			
Cython					
Flask					
gevent					
greenlet				
grequests				
idna					
importlib-metadata		
injector				
isodate					
itsdangerous			
Jinja2					
logging-azure			
MarkupSafe				
msal					
msal-extensions					
portalocker				
pycparser				
PyJWT					
pywin32					
requests				
setuptools				
six						
typing_extensions		
urllib3					
Werkzeug				
zipp					
zope.event				
zope.interface			
flask_login
flask_otp